% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagGGmcmc.R
\name{diagGGmcmc}
\alias{diagGGmcmc}
\title{Plot MCMC Diagnostics}
\usage{
diagGGmcmc(mcmc, filetype = "png", dpi = 200)
}
\arguments{
\item{mcmc}{an \code{mcmc.list} object}
}
\value{
Returns a summary of the \code{mcmc.list} object
}
\description{
This function takes an \code{mcmc.list} object and outputs diagnostic plots to include trace plots, autocorrelation plots, running mean plots, and posterior density plots. It is especially useful to examine if MCMC chains are well-behaved.
}
\examples{
\dontrun{
data(codaSamples)
mcmc=coda::as.mcmc.list(codaSamples)
diagGGmcmc(codaSamples, filetype='png', dpi=200)
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
