% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformedBetaBF.R
\name{InformedBetaBF}
\alias{InformedBetaBF}
\title{Bayes Factor for an Informed Beta Prior Distribution}
\usage{
InformedBetaBF(alpha2, beta2, n, x)
}
\arguments{
\item{alpha2}{The alpha parameter for the informed Beta prior distribution}

\item{beta2}{The beta parameter for the informed Beta prior distribution}

\item{n}{The total number of trials}

\item{x}{The total number of successes}
}
\value{
Returns \eqn{P(M_2)/P(M_1)}, where \eqn{M_1} is the uninformed prior.
}
\description{
This function takes parameters for an informed Beta prior distribution and compares it to the uninformed prior \code{beta(1,1)} and releases a command line output of the Bayes Factor.
}
\examples{
InformedBetaBF(alpha2=2, beta2=5, n=10, x=4)
##Bayes Factor (informed vs uniformed)
##1.442308
##contour plots not shown
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
