% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSOReg.R
\name{LASSOReg}
\alias{LASSOReg}
\title{LASSO Regression using MCMC Sampling}
\usage{
LASSOReg(x, y, xnew = NULL, iter = 3334, burn = 5000, thin = 3,
  alpha = 0.05, parallel = TRUE, phi = NULL, DIC = FALSE)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables}

\item{y}{A vector of data}

\item{xnew}{A vector of predictors for making inference}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}

\item{parallel}{\code{TRUE/FALSE}. Option to use multiple CPUs}

\item{phi}{The constraint value on predictor priors}

\item{DIC}{\code{TRUE/FALSE}. Option to track DIC. Cannot be used in conjunction with \code{parallel} method}
}
\description{
This function takes in a matrix of predictor variables \code{x} and a vector of data \code{y} and runs a LASSO regression. This is especially useful for dealing with multicollinearity issues with the goal of eliminating predictor variables from the model.
}
\examples{
\dontrun{
data(clerical)
y=clerical[,1]
x=as.matrix(clerical[,-1])
xnew=c(5000, 75, 900, 200, 650)
testobj=LASSOReg( x,
            y,
            xnew,
            iter=3334,
            burn=5000,
            thin=3,
            alpha=0.05,
            parallel=TRUE,
            phi=NULL,
            DIC=FALSE)

## Point Estimates
##beta0       beta[1]       beta[2]       beta[3]       beta[4]       beta[5]         sigma           phi         ydash
##67.053598720   0.001220227   0.108111808   0.012961810  -0.043823459   0.055245865  11.225304473   0.500000000 120.176385903
##Interval Estimates
##[[1]]
##lower        upper
##beta0   49.315400000  84.33490000
##beta[1] -0.000400664   0.00300408
##beta[2]  0.017713400   0.19708500
##beta[3] -0.005275380   0.02977720
##beta[4] -0.078546200  -0.00877705
##beta[5]  0.034261800   0.07875780
##sigma    9.130320000  13.81670000
##phi      0.500000000   0.50000000
##ydash   97.900300000 141.97700000
##attr(,"Probability")
##[1] 0.94991
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
