% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinQuadReg.R
\name{LinQuadReg}
\alias{LinQuadReg}
\title{Linear vs Quadratic Regression Comparison for One Predictor Variable}
\usage{
LinQuadReg(x, y, xnew = NULL, iter = 3334, burn = 5000, thin = 3,
  alpha = 0.05)
}
\arguments{
\item{x}{A vector of predictor variables}

\item{y}{A vector of data}

\item{xnew}{A value of a predictor for making inference}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}
}
\description{
This function takes in a vector of predictor variables \code{x} and a vector of data \code{y} and creates a quadratic model and a simple linear model using MCMC sampling. For the Bayesian inference, this method uses uninformative prior distributions. The point and interval estimates are taken from the posterior distribution from the samples. The models are then plotted over the data and DIC is calculated to determine goodness of fit
}
\examples{
\dontrun{
data("electric")
y=electric[,3]
x=electric[,1]
LinQuadReg(  x,
             y,
             xnew=NULL,
             iter=3334,
             burn=5000,
             thin=3,
             alpha=0.05)
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
