% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeReg.R
\name{RidgeReg}
\alias{RidgeReg}
\title{Ridge Regression using MCMC Sampling}
\usage{
RidgeReg(x, y, xnew = NULL, iter = 3334, burn = 5000, thin = 3,
  alpha = 0.05, parallel = TRUE, phi = NULL, DIC = FALSE)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables}

\item{y}{A vector of data}

\item{xnew}{A vector of predictors for making inference}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}

\item{parallel}{\code{TRUE/FALSE}. Option to use multiple CPUs}

\item{phi}{The constraint value on predictor priors}

\item{DIC}{\code{TRUE/FALSE}. Option to track DIC. Cannot be used in conjunction with \code{parallel} method}
}
\description{
This function takes in a matrix of predictor variables \code{x} and a vector of data \code{y} and runs a Ridge regression. This is especially useful for dealing with multicollinearity issues and wanting to keep all predictor variables in the model.
}
\examples{
\dontrun{
data(clerical)
y=clerical[,1]
x=as.matrix(clerical[,-1])
xnew=c(5000, 75, 900, 200, 650)
testobj=RidgeReg( x,
             y,
             xnew,
             iter=3334,
             burn=5000,
             thin=3,
             alpha=0.05,
             parallel=TRUE,
             phi=0.5,
             DIC=FALSE)
# commmand line output
##Point Estimates
##beta0       beta[1]       beta[2]       beta[3]       beta[4]       beta[5]         sigma           phi         ydash
##66.374917107   0.001239869   0.114532661   0.012663400  -0.045415560   0.056192030  11.218459152   0.500000000 120.112178034
##Interval Estimates
##[[1]]
##lower        upper
##beta0   49.04310000  83.39460000
##beta[1] -0.00055239   0.00294774
##beta[2]  0.02783660   0.20953100
##beta[3] -0.00431864   0.03036180
##beta[4] -0.08085720  -0.01009940
##beta[5]  0.03455220   0.07838760
##sigma    8.97739000  13.44550000
##phi      0.50000000   0.50000000
##ydash   97.10990000 140.40900000
##attr(,"Probability")
##[1] 0.94991
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
