% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoTTest.R
\name{two.t.test}
\alias{two.t.test}
\title{Two-Sample Bayesian T-Test}
\usage{
two.t.test(y1, y2, iter = 3334, burn = 5000, thin = 3,
  alpha = 0.05, parallel = FALSE)
}
\arguments{
\item{y1, y2}{A vector of data}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}

\item{parallel}{\code{TRUE/FALSE}. Option to use multiple CPUs}
}
\description{
This function takes in two vectors of data \code{y1,y2} and performs a t-test using MCMC sampling. For the Bayesian inference, this method uses uninformative prior distributions. The point and interval estimates are taken from the posterior distribution from the samples. The function also generates an interactive 3D contour plot of the two sample densities.
}
\examples{
\dontrun{
data("reading")
y1=reading$Grade[reading$Method == 'c']
y2=reading$Grade[reading$Method == 't']
testobj=two.t.test(y1,
                   y2,
                   iter=3334,
                   burn=5000,
                   thin=3,
                   alpha=0.05,
                   parallel=TRUE)
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
