% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UninformedMLR.R
\name{UninformedMLR}
\alias{UninformedMLR}
\title{Multiple Regression using MCMC Sampling}
\usage{
UninformedMLR(x, y, xnew = NULL, iter = 3334, burn = 5000,
  thin = 3, alpha = 0.05, parallel = TRUE, DIC = FALSE)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables}

\item{y}{A vector of data}

\item{xnew}{A vector of predictors for making inference. Must be equal length to \code{x}.}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}

\item{parallel}{\code{TRUE/FALSE}. Option to use multiple CPUs}

\item{DIC}{\code{TRUE/FALSE}. Option to track DIC. Cannot be used in conjunction with \code{parallel} method}
}
\description{
This function takes in a matrix of predictor variables \code{x} and a vector of data \code{y} and runs a multiple regression using uninformative priors. It will require many iterations for the autocorrelation to shrink if the predictors have high multicollinearity. Before using this function, it is suggested to use the \code{corrcheck()} function to determine whether it is more appropriate to use the \code{RidgeReg()} function or the \code{LASSOReg()} functions.
}
\examples{
\dontrun{
data("iqsize")
x=as.matrix(iqsize[,c('Brain','Height','Weight')])
y=iqsize[,'PIQ']
testobj=UninformedMLR(x,
                      y,
                      iter=3334,
                      burn=5000,
                      thin=3,
                      alpha=0.05,
                      parallel=TRUE,
                      DIC=FALSE)
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
