% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UninformedSLR.R
\name{UninformedSLR}
\alias{UninformedSLR}
\title{Simple Linear Regression with Uninformative Priors}
\usage{
UninformedSLR(x, y, xnew = NULL, iter = 3334, burn = 5000,
  thin = 3, alpha = 0.05, parallel = TRUE, DIC = FALSE)
}
\arguments{
\item{x}{A vector of predictor variables}

\item{y}{A vector of data}

\item{xnew}{A value of a predictor for making inference}

\item{iter}{The number of iterations per chain}

\item{burn}{The burn in period for chains}

\item{thin}{The thinning interval for monitoring}

\item{alpha}{For setting the \eqn{(1-\alpha)} Highest Density Interval}

\item{parallel}{\code{TRUE/FALSE}. Option to use multiple CPUs}

\item{DIC}{\code{TRUE/FALSE}. Option to track DIC. Cannot be used in conjunction with \code{parallel} method}
}
\description{
This function takes in a vector of predictor variables \code{x} and a vector of data \code{y} and creates a simple linear model using MCMC sampling. For the Bayesian inference, this method uses uninformative prior distributions. The point and interval estimates are taken from the posterior distribution from the samples.
}
\examples{
\dontrun{
data("electric")
y=electric[,3]
x=electric[,1]
testobj=UninformedSLR(x,
                      y,
                      xnew=1700,
                      iter=3334,
                      burn=5000,
                      thin=3,
                      alpha=0.05,
                      parallel=TRUE,
                      DIC=TRUE)
}
}
\author{
Jeff Nivitanont, \email{jeffniv@ou.edu}
}
